requirements.txt
Django>=3.2
cryptography>=3.4
boto3>=1.18
python-dotenv
